@@ -3,35 +3,56 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "node app.js",
+    "dev": "nodemon app.js",
     "build": "ng build",
+    "prod": "ng build --prod",
+    "prod:watch": "ng build --prod --watch",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.6",
-    "@angular/common": "~10.1.6",
-    "@angular/compiler": "~10.1.6",
-    "@angular/core": "~10.1.6",
-    "@angular/forms": "~10.1.6",
-    "@angular/platform-browser": "~10.1.6",
-    "@angular/platform-browser-dynamic": "~10.1.6",
-    "@angular/router": "~10.1.6",
-    "ngx-spinner": "^10.0.1",
-    "ngx-toastr": "^13.1.0",
+    "@angular/animations": "^10.2.3",
+    "@angular/cdk": "^10.2.7",
+    "@angular/common": "~10.2.0",
+    "@angular/compiler": "~10.2.0",
+    "@angular/core": "~10.2.0",
+    "@angular/forms": "~10.2.0",
+    "@angular/material": "^10.2.7",
+    "@angular/platform-browser": "~10.2.0",
+    "@angular/platform-browser-dynamic": "~10.2.0",
+    "@angular/router": "~10.2.0",
+    "@bigcommerce/big-design": "^0.25.0",
+    "bootstrap": "^4.5.3",
+    "cookie-parser": "^1.4.5",
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "got": "^11.8.0",
+    "jquery": "^3.5.1",
+    "morgan": "^1.10.0",
+    "ngx-toastr": "^13.2.0",
+    "ngx-ui-loader": "^10.0.0",
+    "node-bigcommerce": "^4.1.0",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "rxjs": "~6.6.0",
+    "styled-components": "^5.2.1",
     "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1001.7",
-    "@angular/cli": "~10.1.7",
-    "@angular/compiler-cli": "~10.1.6",
-    "@types/node": "^12.11.1",
+    "@angular-devkit/build-angular": "~0.1002.0",
+    "@angular/cli": "~10.2.0",
+    "@angular/compiler-cli": "~10.2.0",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.11.1",
+    "@types/react": "^16.9.17",
+    "@types/react-dom": "^16.9.4",
+    "@types/styled-components": "^5.1.5",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
@@ -40,6 +61,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
+    "nodemon": "^2.0.6",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
