@@ -1,9 +1,15 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './components/home/home.component';
 
 const routes: Routes = [
-  { path: '', component: HomeComponent }
+  {
+    path: "",
+    loadChildren: () => import("./components/components.module").then((m) => m.ComponentsModule),
+  },
+  {
+    path: "**",
+    redirectTo: ""
+  }
 ];
 
 @NgModule({
